requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Mail==0.9.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
google-generativeai==0.3.2
Werkzeug==3.0.1
SQLAlchemy==2.0.23
